requirements.txt
bs4
cryptography
PyYAML
